+++ src/vue/forms/ChangePasswordForm.vue
@@ -45,6 +45,21 @@
         />
       </div>
     </div>
+    <div
+      v-if="isTotpEnabled"
+      class="app__form-row"
+    >
+      <div class="app__form-field">
+        <input-field
+          v-model="form.tfaCode"
+          @blur="touchField('form.tfaCode')"
+          id="change-password-tfa-code"
+          :error-message="getFieldErrorMessage('form.tfaCode')"
+          :label="'change-password-form.tfa-code-lbl' | globalize"
+          :disabled="formMixin.isDisabled"
+        />
+      </div>
+    </div>
 
     <div class="app__form-actions">
       <form-confirmation
@@ -68,7 +83,7 @@
 <script>
 import FormMixin from '@/vue/mixins/form.mixin'
 
-import { required, password, sameAs } from '@validators'
+import { required, requiredIf, password, sameAs } from '@validators'
 
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import { Bus } from '@/js/helpers/event-bus'
@@ -88,6 +103,7 @@ export default {
       currentPassword: '',
       newPassword: '',
       confirmPassword: '',
+      tfaCode: '',
     },
   }),
   validations: {
@@ -99,11 +115,15 @@ export default {
         password,
         sameAsPassword: sameAs(function () { return this.form.newPassword }),
       },
+      tfaCode: {
+        required: requiredIf(function () { return this.isTotpEnabled }),
+      },
     },
   },
   computed: {
     ...mapGetters({
       wallet: vuexTypes.wallet,
+      isTotpEnabled: vuexTypes.isTotpEnabled,
       accountId: vuexTypes.accountId,
     }),
   },
@@ -120,6 +140,8 @@ export default {
         await Sdk.api.wallets.changePassword(this.form.newPassword)
       } catch (e) {
         if (e instanceof errors.TFARequiredError) {
+          // To change password we should verify password factor first.
+          // To do that we need to provide TFARequiredError instance.
           await this.retryPasswordChange(e)
         } else {
           ErrorHandler.process(e)
@@ -127,22 +149,73 @@ export default {
       }
       this.enableForm()
     },
+
     async retryPasswordChange (tfaError) {
       try {
         await Sdk.api.factors.verifyPasswordFactorAndRetry(tfaError,
           this.form.currentPassword
         )
+      } catch (e) {
+        // If 2FA is enabled we should verify TOTP factor
+        // (using TFARequiredError instance).
+        if (e instanceof errors.TFARequiredError) {
+          try {
+            await Sdk.api.factors.verifyTotpFactorAndRetry(e,
+              this.form.tfaCode
+            )
+          } catch (e) {
+            // FIXME: We need to verify password factor again after
+            // verifying 2FA factor.
+            if (e instanceof errors.TFARequiredError) {
+              await Sdk.api.factors.verifyPasswordFactorAndRetry(e,
+                this.form.currentPassword
+              )
+            } else {
+              // If verifyTotpFactor threw an error different from
+              // TFARequiredError, there must be wrong 2FA code provided.
+              ErrorHandler.process(e, 'change-password-form.wrong-code-err')
+              return
+            }
+          }
+        } else {
+          // If verifyPasswordFactor threw an error different from
+          // TFARequiredError, there must be wrong password provided.
+          ErrorHandler.process(e, 'change-password-form.wrong-password-err')
+          return
+        }
+      }
+
+      // Load new wallet after successful password change.
+      try {
         await this.useNewWallet()
         Bus.success('change-password-form.password-changed-msg')
       } catch (e) {
-        ErrorHandler.process(e, 'change-password-form.wrong-password-err')
+        ErrorHandler.process(e)
       }
     },
+
     async useNewWallet () {
-      const newWallet = await Sdk.api.wallets.get(
-        this.wallet.email,
-        this.form.newPassword
-      )
+      let newWallet
+      try {
+        newWallet = await Sdk.api.wallets.get(
+          this.wallet.email,
+          this.form.newPassword
+        )
+      } catch (e) {
+        // If 2FA is enabled we should verify TOTP factor
+        // to get a user's wallet.
+        if (e instanceof errors.TFARequiredError) {
+          await Sdk.api.factors.verifyTotpFactor(e,
+            this.form.tfaCode
+          )
+          newWallet = await Sdk.api.wallets.get(
+            this.wallet.email,
+            this.form.newPassword
+          )
+        } else {
+          throw e
+        }
+      }
       Api.useWallet(newWallet)
       Sdk.sdk.useWallet(newWallet)
       this.storeWallet(newWallet)